Separate functions for line in utils
